Folders for Blue and C robots
