+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Blue_Bot/BlueWheelTWB.java
@@ -1,19 +1,26 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Blue_Bot;
 
 import android.annotation.SuppressLint;
 
+import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
+
+import org.firstinspires.ftc.teamcode.ArmServo;
+import org.firstinspires.ftc.teamcode.DatalogTWB;
+import org.firstinspires.ftc.teamcode.PiecewiseFunction;
+import org.firstinspires.ftc.teamcode.TwoWheelBalanceController;
+
+import java.util.Locale;
 
 /**
  * Blue Wheeled Two Wheel Balancing Robot Class, with Arm.
  *  All of the robot unique constant are (should be) defined here
  */
-public class BlueWheelTWB {
-    private final TwoWheelBalanceController TWBController;
-
+public class BlueWheelTWB extends TwoWheelBalanceController{
     private final Servo clawServo;
     private boolean ClawIsClosed = false; //Claw boolean
     private final double CLAWCLOSE = 1.0; // servo value for closed claw
@@ -28,24 +35,24 @@ public class BlueWheelTWB {
 
     // PieceWise linear curve member for pitch angle vs arm angle
     final private PiecewiseFunction pitchAngVec = new PiecewiseFunction();
-    private DatalogTWB datalogTWB; // datalog for full recording
-    private boolean writeDatalog = false; // default is no log.  call method to write.
-    public double MMPLoop = 8.0;
-    public double DEGPLoop = 1.0;
+
+
     /**
      * TWB Constructor.  Call once.
       */
     public BlueWheelTWB(HardwareMap hardwareMap) {
-
+        super(hardwareMap, 300.0,
+                1.75619, 0.45, 0.0, 0.05, 7, 3);
         // The distance between the blue wheels is 300 mm
         // REVSPUR40PPR = 1120; // REV Core Hex Motor Pulses per Revolution at output shaft
         // COUNTS_PER_REV    = 2048.0 ;    // CUI ATM103 Encoder at most PPR
         // WHEELDIA = 203.0; // 8 inch wheel diameter (mm)
         // TICKSPERMM = (1120)/(203*Math.PI) = 1.75619; // REV SPUR 40:1, 8in wheels
         // Yaw PID terms: kp 0.45, ki 0.12, kd 0.05
-        TWBController = new TwoWheelBalanceController(hardwareMap, 300.0,
-                1.75619, 0.45, 0.0, 0.05, 7, 3,
-                TwoWheelBalanceController.Robot.Blue);
+
+        this.imu.initialize(new IMU.Parameters(new RevHubOrientationOnRobot(
+                RevHubOrientationOnRobot.LogoFacingDirection.UP,
+                RevHubOrientationOnRobot.UsbFacingDirection.BACKWARD)));
 
         // These are the state terms for a two wheel balancing robot
         // Tune these using the DOE (Design of Experiments) opmode.
@@ -54,9 +61,14 @@ public class BlueWheelTWB {
         //                            Kpos        Kvelo       Kpitch       KpitchRate
         //TWBController.setBalanceTerms(0.020,0.018,-0.58,-0.028);
         //                                  0.016       0.015       -0.58           -0.025
-        TWBController.setBalanceTerms(0.001785,0.00125,-0.04845,-0.002125);
+        setBalanceTerms(0.001785,0.00125,-0.04845,-0.002125);
 
-        TWBController.setDriveMotors(false, true, false);
+        setMMPLoop(8.0);
+        setDEGPLoop(0.5);
+
+        setDriveMotors(false, true, false);
+
+        setVerticalCM(92.0);  // mm
 
         // Initialize the arm class
         // Determine servo values for two angle using the ServoTester opmode
@@ -103,41 +115,32 @@ public class BlueWheelTWB {
      */
     public void auto_right_loop() {
         // Check which way the robot is leaning and rotate the arm so that it will self-right
-        if (TWBController.getNewPitch() > 0.0) theArm.setArmAngle(ARMMAX);
+        if (getNewPitch() > 0.0) theArm.setArmAngle(ARMMAX);
         else theArm.setArmAngle(-125.0);  // -140 drives IMU pitch past zero, so using less arm angle
 
-        theArm.updateArm(TWBController.getDeltaTime()); // This will make the arm move
+        theArm.updateArm(getDeltaTime()); // This will make the arm move
     }
 
     /**
      * Start calls the TWB controller start
      */
     public void start() {
-        TWBController.start();
+        super.start();
         clawTimer.reset();
+        imuYawPitchReset();
     }
     /**
      * TWB Main Loop method.  Call repeatedly while running. Contains balance control logic.
      * Teleoperated inputs are removed from this method, so it can be called in autonomous.
       */
-    @SuppressLint("DefaultLocale")
-    public void loop(OpMode theOpmode) {
+    public void loopBlue(OpMode theOpmode) {
 
-        TWBController.setArmPitchTarget(pitchAngVec.getY(theArm.getAngle()));
+        setArmPitchTarget(pitchAngVec.getY(theArm.getAngle()));
 
-        TWBController.loop(theOpmode);
+        loop(theOpmode);
 
-        theArm.updateArm(TWBController.getDeltaTime()); // This will make the arm move
+        theArm.updateArm(getDeltaTime()); // This will make the arm move
 
-        if (writeDatalog) {
-            datalogTWB.logPosPitch(TWBController.getPos(), TWBController.getPosTarget(),
-                    TWBController.getVelocity(), TWBController.getVeloTarget(),TWBController.getPitch(),
-                    TWBController.getPitchTarget(), TWBController.getPitchRate(),
-                    TWBController.getYaw(),TWBController.getYawTarget(),
-                    TWBController.getPositionVolts(),TWBController.getPitchVolts(),
-                    TWBController.getDeltaTime());
-            datalogTWB.writeLineTWB();
-        }
     }
 
     /**
@@ -185,76 +188,19 @@ public class BlueWheelTWB {
                 ClawIsClosed = !ClawIsClosed;
             }
         }
-
     }
 
-    /**
-     * TWB method to provide user control of turning the robot.
-     * @param deltaAngle a value from -1 to 1 in radians
-     */
-    public void turn_teleop(double deltaAngle) {
-        // Robot Turning:
-        // The right joystick turns the robot by adjusting the yaw PID turn setpoint
-        TWBController.setYawTarget(TWBController.getYawTarget() - deltaAngle );
-    }
-
-    /**
-     * TWB method translates the robot at the current angle by setting Position, Velocity & Pitch Targets.
-     * @param forward value from -1 to 1 that is the forward or backward amount
-     * @param degPerLoop robot pitch degrees per loop, multiplied by forward (6 is good)
-     * @param mmPerLoop robot translation in mm per loop, multiplied by forward (7 is good)
-     */
-    public void translateDrive(double forward, double mmPerLoop, double degPerLoop) {
-
-        // add some pitch to get it moving (degPerLoop of 6 results in gentle movement)
-        TWBController.setAutoPitchTarget(forward * degPerLoop);
-
-        // Update posTarget (mm) Note: this value * 50 = mm per second
-        // mmPerLoop of 7 results in a gentle speed
-        TWBController.setPosTarget( TWBController.getPosTarget() - forward * mmPerLoop );
-
-        // Update the velocity target (mm/sec)
-        //TWBController.setVeloTarget( -forward*(mmPerLoop/0.020));
-        //TWBController.setVeloTarget( -forward*(mmPerLoop/TWBController.getDeltaTime()));
-    }
-    @SuppressLint("DefaultLocale")
     public void writeTelemetry(OpMode om) {
-        om.telemetry.addLine(String.format("s Position Target %.1f ,Current %.1f (mm)",
-                TWBController.getPosTarget(),TWBController.getPos()));
-        om.telemetry.addLine(String.format("s Velocity Target %.1f ,Current %.1f (mm/sec)",
-                TWBController.getVeloTarget(),TWBController.getVelocity()));
-        om.telemetry.addLine(String.format("Pitch Target %.1f ,Current %.1f (degrees)",
-                TWBController.getPitchTarget(),TWBController.getPitch()));
-        om.telemetry.addLine(String.format("Arm Angle Target %.1f ,Current %.1f (degrees)",
+        om.telemetry.addLine(String.format(Locale.US,"s Position Target %.1f ,Current %.1f (mm)",
+                getPosTarget(),getPos()));
+        om.telemetry.addLine(String.format(Locale.US,"s Velocity Target %.1f ,Current %.1f (mm/sec)",
+                getVeloTarget(),getVelocity()));
+        om.telemetry.addLine(String.format(Locale.US,"Pitch Target %.1f ,Current %.1f (degrees)",
+                getPitchTarget(),getPitch()));
+        om.telemetry.addLine(String.format(Locale.US,"Arm Angle Target %.1f ,Current %.1f (degrees)",
                 theArm.getTargetAngle(),theArm.getAngle()));
-        //om.telemetry.addData("Current Voltage   ",TWBController.getCurrentVoltage());
     }
-
     public void closeClaw() {clawServo.setPosition(CLAWCLOSE);}
     public void openClaw() {clawServo.setPosition(CLAWOPEN);}
-    public double getKpos() {return TWBController.getKpos();}
-    public double getKpitch() {return TWBController.getKpitch();}
-    public double getKvelo() {return TWBController.getKvelo();}
-    public double getKpitchRate() {return TWBController.getKpitchRate();}
-    public void setKpos(double k) {TWBController.setKpos(k);}
-    public void setKpitch(double k) {TWBController.setKpitch(k);}
-    public void setKpitchRate(double k) {TWBController.setKpitchRate(k);}
-    public void setKvelo(double k) {TWBController.setKvelo(k);}
-    public void setAutoPitchTarget(double target) {TWBController.setAutoPitchTarget(target);}
-    public void setPosTarget(double pos) {TWBController.setPosTarget(pos);}
-    public void setVeloTarget(double velo) {TWBController.setVeloTarget(velo);}
-    public double getPos() {return TWBController.getPos();}
-    public double getVelocity() {return TWBController.getVelocity();}
-    public double getPitch() {return TWBController.getPitch();}
-    public double getPitchTarget() {return TWBController.getPitchTarget();}
-    public double getDeltaTime() {return TWBController.getDeltaTime();}
-    public void writeDatalog(String LogName) {
-        this.writeDatalog=true;
-        datalogTWB = new DatalogTWB();
-        datalogTWB.init(LogName);
-    }
-    public void setYawTarget(double yaw) {TWBController.setYawTarget(yaw);}
-    public double getYawTarget() {return TWBController.getYawTarget();}
-    public double getYaw() {return TWBController.getYaw();}
-    public void imuReset() {TWBController.imuYawReset();}
+
 }
