+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Blue_Bot/Blue_back_n_forth_DOE.java
@@ -1,5 +1,5 @@
 
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.Blue_Bot;
 
 import android.annotation.SuppressLint;
 
@@ -8,7 +8,6 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.BlueWheelTWB;
 import org.firstinspires.ftc.teamcode.Datalogger;
 import org.firstinspires.ftc.teamcode.TWBMove;
 import org.firstinspires.ftc.teamcode.Term;
@@ -150,7 +149,7 @@ public class Blue_back_n_forth_DOE extends OpMode {
                 break;
         }
 
-        twb.loop(this);  // MAIN CONTROL SYSTEM
+        twb.loopBlue(this);  // MAIN CONTROL SYSTEM
 
         telemetry.addLine(String.format("EXPERIMENT %d ,OF TOTAL %d", count, NEXPERIMENTS));
         telemetry.addData("State", state);
