+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/C_Bot/C_DriveSimple.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.C_Bot;
 
 import android.annotation.SuppressLint;
